requirements.txt
pandas==2.2.3
numpy==2.1.3
matplotlib==3.10.0
seaborn==0.13.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.39
python-dotenv==1.1.0
requests==2.32.3
